+++ src/analyzers/concept/poetry-club-door-policy/index.ts
@@ -3,3 +3,5 @@ import { ExemplarAnalyzer } from '../__exemplar'
 export class PoetryClubDoorPolicyAnalyzer extends ExemplarAnalyzer {
   // TODO: implement actual analyzer
 }
+
+export default PoetryClubDoorPolicyAnalyzer